+++ src/extensions/askUserForParam.ts
@@ -187,22 +187,22 @@ export async function askForParam(args: {
         });
         valid = checkResponse(response, paramSpec);
         break;
-      case ParamType.MULTISELECT:
+      case ParamType.MULTISELECT: {
+        const defaultValues = paramSpec.default ? paramSpec.default.split(",") : undefined;
         response = (
           await checkbox<string>({
-            default: paramSpec.default
-              ? paramSpec.default.split(",").map((def) => {
-                  return getInquirerDefault(_.get(paramSpec, "options", []), def);
-                })
-              : undefined,
             message:
               "Which options do you want enabled for this parameter? " +
               "Press Space to select, then Enter to confirm your choices. ",
-            choices: convertExtensionOptionToLabeledList(paramSpec.options as ParamOption[]),
+            choices: convertExtensionOptionToLabeledList(
+              paramSpec.options as ParamOption[],
+              defaultValues,
+            ),
           })
         ).join(",");
         valid = checkResponse(response, paramSpec);
         break;
+      }
       case ParamType.SECRET:
         do {
           secretLocations = await promptSecretLocations(paramSpec);
@@ -223,7 +223,7 @@ export async function askForParam(args: {
       default:
         // Default to ParamType.STRING
         response = await input({
-          default: paramSpec.default,
+          default: paramSpec.default ?? (paramSpec.required ? undefined : ""),
           message: `Enter a value for ${label}:`,
         });
         valid = checkResponse(response, paramSpec);
@@ -414,5 +414,5 @@ async function addNewSecretVersion(
  */
 export function getInquirerDefault(options: ParamOption[], def: string): string {
   const defaultOption = options.find((o) => o.value === def);
-  return defaultOption ? defaultOption.label || defaultOption.value : "";
+  return defaultOption ? defaultOption.value : def;
 }